+++ client/src/component/employee/Payroll.js
@@ -1,2 +1,11 @@
 import React from 'react'
 
+const Payroll = () =>{
+    return (
+        <div>
+            Payroll
+        </div>
+    )
+}
+
+export default Payroll
